svg/painting/parsing/image-rendering-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing image-rendering with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ImageRenderingProperty" rel="help"/><h:meta content="image-rendering supports the full grammar 'auto | optimizeQuality | optimizeSpeed'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("image-rendering", "auto");
test_valid_value("image-rendering", "optimizequality");
test_valid_value("image-rendering", "optimizespeed");

  ]]></script></svg>
